Revert "Test a rule from the rule manager service with CAPI content"
