+++ apps/checker/conf/application.conf
@@ -8,9 +8,6 @@ play.assets {
   urlPrefix = "/assets"
 }
 
-# Permit large payloads
-play.http.parser.maxMemoryBuffer=4MB
-
 typerighter {
   sheetRange="A:G"
   defaultAwsProfile = "composer"
