+++ apps/checker/test/scala/model/TextRangeTest.scala
