+++ apps/common-lib/src/main/scala/com/gu/typerighter/lib/CommonConfig.scala
@@ -25,8 +25,6 @@ abstract class CommonConfig(
     val ws: WSClient
 ) extends Loggable {
   val serviceName: String
-  val capiApiKey = playConfig.get[String]("capi.apiKey")
-
   val permissionsBucket =
     playConfig.getOptional[String]("permissions.bucket").getOrElse("permissions-cache")
