+++ apps/rule-manager/app/AppComponents.scala
@@ -8,8 +8,7 @@ import com.amazonaws.auth.AWSCredentialsProvider
 import com.gu.AwsIdentity
 import com.gu.AppIdentity
 import com.gu.DevIdentity
-import com.gu.contentapi.client.GuardianContentClient
-import com.gu.typerighter.lib.{HMACClient, ContentClient}
+import com.gu.typerighter.lib.HMACClient
 import com.gu.typerighter.rules.BucketRuleResource
 import router.Routes
 import db.DB
@@ -55,9 +54,7 @@ class AppComponents(
   val hmacClient = new HMACClient(stage, secretKey = config.hmacSecrets.head)
   val sheetsRuleResource = new SheetsRuleResource(config.credentials, config.spreadsheetId)
   val bucketRuleResource = new BucketRuleResource(s3Client, typerighterBucket, stage)
-  val guardianContentClient = GuardianContentClient(config.capiApiKey)
-  val contentClient = new ContentClient(guardianContentClient)
-  val ruleTesting = new RuleTesting(wsClient, hmacClient, contentClient, config.checkerServiceUrl)
+  val ruleTesting = new RuleTesting(wsClient, hmacClient, config.checkerServiceUrl)
 
   val homeController = new HomeController(
     controllerComponents,
